+++ test/actions.test.js
@@ -6,50 +6,6 @@ beforeEach(() => {
   document.body.innerHTML = ""
 })
 
-test("slices", done => {
-  const actions = app({
-    state: {
-      foo: {
-        bar: {
-          baz: "minimal baz"
-        }
-      }
-    },
-    view: state =>
-      h(
-        "div",
-        {
-          oncreate() {
-            expect(document.body.innerHTML).toBe(
-              `<div>only the baz will do</div>`
-            )
-            done()
-          }
-        },
-        state.foo.bar.baz
-      ),
-    actions: {
-      foo: {
-        bar: {
-          baz: state => data => {
-            expect(state).toEqual({ baz: "minimal baz" })
-            expect(data).toBe("foo.bar.baz")
-            return { baz: "only the baz will do" }
-          }
-        }
-      },
-      fizz: {
-        buzz: {
-          fizzbuzz: () => ({ fizzbuzz: "fizzbuz" })
-        }
-      }
-    }
-  })
-
-  actions.foo.bar.baz("foo.bar.baz")
-  actions.fizz.buzz.fizzbuzz()
-})
-
 test("sync updates", done => {
   app({
     state: {
