--- test/ts/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-  "strict": true,
-  "noEmit": true,
-  "jsx": "react",
-  "jsxFactory": "h",
-  "baseUrl": "../..",
-    "paths": {
-      "classwrap": ["src", "hyperapp.d.ts"]
-    }
-  },
-  "include": [
-    "*.ts",
-    "*.tsx",
-    "**/*.ts",
-    "**/*.tsx"
-  ]
-}
